backend/Neotube/requirements.txt
django ~= 3.1.6
djangorestframework ~= 3.12.2
djangorestframework-jwt ~= 1.11.0
django-cors-headers
json
httpie
